backend/src/main/java/com/example/moneykeeper/controller/MainController.java
package com.example.moneykeeper.controller;

import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;

import java.security.Principal;

@RestController
@RequestMapping("user")
public class MainController {
    @GetMapping
    public ResponseEntity<String> getUsername(Principal principal) {
        return ResponseEntity.ok()
                .contentType(MediaType.APPLICATION_JSON)
                .body(principal.getName());
    }
}
