backend/src/test/java/com/example/moneykeeper/controller/MainControllerTest.java
package com.example.moneykeeper.controller;

import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.mockito.junit.jupiter.MockitoExtension;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;

import java.security.Principal;

import static org.junit.jupiter.api.Assertions.*;
import static org.mockito.Mockito.when;

@ExtendWith(MockitoExtension.class)
class MainControllerTest {

    @Mock
    Principal principal;

    @InjectMocks
    MainController controller;

    @Test
    void userAccess_ReturnsValidResponse() {
        //given
        String username = "username";
        when(principal.getName()).thenReturn(username);

        // when
        ResponseEntity<String> response = controller.getUsername(principal);

        // then
        assertNotNull(response);
        assertEquals(HttpStatus.OK, response.getStatusCode());
        assertEquals(MediaType.APPLICATION_JSON, response.getHeaders().getContentType());
        assertEquals(username, response.getBody());
    }
}